+++ R/bootstrap_quanta.R
@@ -1,4 +1,9 @@
-#' Compute boostrap confidence interval
+#' Compute Boostrap Confidence Interval
+#'
+#'
+#' This function employs the Monte Carlo bootstrap method to determine the
+#' confidence interval that delimits the range of most possible quantums for
+#' a given set of measurements.
 #'
 #' @param set a numeric vector of measurements from which quantum and bootstrap
 #' should be computed
@@ -6,6 +11,8 @@
 #' @param n_samples number of bootstrap samples
 #' @param sample_portion The portion (<1) of the original data that should
 #' be used for each sample
+#' @param unround TRUE/FALSE, declares if artificial noise should be added to
+#' the measurements (T) or if they should be taken as is (F), defaults to TRUE.
 #' @param params Constants used in the quantum calculation,
 #' see getOption("CONSTANTS_QUANTOGRAM")
 #'
@@ -22,6 +29,7 @@ bootstrap_quanta <- function(set = NULL,
                              sample_portion = 0.75,
                              confint = 0.05,
                              n_samples = 500,
+                             unround = TRUE,
                              params = getOption("CONSTANTS_QUANTOGRAM")) {
 
   if (!is.numeric(sample_portion)) {
@@ -30,7 +38,7 @@ bootstrap_quanta <- function(set = NULL,
     stop("sample_portion should be 1 or less")
   }
 
-  original_quanta <- get_quantum(set, fofq = TRUE)
+  original_quanta <- get_quantum(set, fofq = TRUE, unround = unround)
   results <- rep(list(NA), n_samples)
 
   decrease <- length(set) - floor(length(set) * sample_portion)
@@ -39,7 +47,9 @@ bootstrap_quanta <- function(set = NULL,
 
   for (n in 1:n_samples) {
     sample <- sample_bootstrap(set, decrease_by = decrease)
-    sample_quantum <- cbind(sample_nr = n, get_quantum(sample, fofq = TRUE))
+    sample_quantum <- cbind(sample_nr = n, get_quantum(sample,
+                                                       fofq = TRUE,
+                                                       unround = unround))
     number_of_peaks <- nrow(sample_quantum)
     if (number_of_peaks > 1) {
       multiple_peaks[n] <- TRUE