+++ NEWS.md
@@ -1,3 +1,11 @@
+# quantatools 0.0.3 _2021-06-26_
+
+* Reduced functions to core (Kendalls Cosine Quantogram related, including
+plots and bootstrap)
+* Updated Vignette
+* Updated Readme 
+* Updated Documentation
+
 # quantatools 0.0.2 _2021-04-15_
 
 * Added a `NEWS.md` file to track changes to the package.